backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
httpx==0.27.2
pydantic==2.9.2
yfinance>=0.2.54
openai==1.54.3
finnhub-python==2.4.20
scipy==1.14.1
websockets==13.1
